pass wifi payload to uart

+++ include/wifi_headers/protocol_manager.h
@@ -7,7 +7,7 @@
 
 // Callback chiamata quando arriva un comando valido dal computer.
 // Contiene gli angoli dei servomotori (1-5 valori, range 0-270).
-using ServoCommandCallback = std::function<void(const std::vector<uint16_t>& angles)>;
+using ServoCommandCallback = std::function<void(const std::vector<float>& angles)>;
 
 class ProtocolManager {
 public:
